+++ config/config.php
@@ -3,10 +3,16 @@
 use GuzzleHttp\RequestOptions;
 
 return [
+    'default_client' => 'default',
+
     'default_config' => [
         RequestOptions::TIMEOUT => 5,
         RequestOptions::CONNECT_TIMEOUT => 1,
         RequestOptions::HTTP_ERRORS => true,
         RequestOptions::ALLOW_REDIRECTS => true,
+    ],
+
+    'clients' => [
+        'default' => [],
     ],
 ];